+++ includes/rcapi/requests/class-wp-relacoof-btoc-get-configuration.php
@@ -33,6 +33,7 @@ class WP_Relacoof_B2C_Get_Configuration extends WP_Relais_Colis_Request {
     const MODULE_VERSION = 'moduleVersion';
     const CMS_NAME = 'cmsName';
     const CMS_VERSION = 'cmsVersion';
+    const HANDLE_NEW_ERRORS = 'handleNewError';
 
     private $mandatory_params = array(
         self::ACTIVATION_KEY,
@@ -70,6 +71,7 @@ class WP_Relacoof_B2C_Get_Configuration extends WP_Relais_Colis_Request {
             self::MODULE_VERSION => Relais_Colis_Woocommerce_Loader::instance()->get_version(),
             self::CMS_NAME => Relais_Colis_Woocommerce_Loader::CMS_WORDPRESS,
             self::CMS_VERSION => $wp_version,
+            self::HANDLE_NEW_ERRORS => true,
         );
 
         // No params